Remove URIs from HostJson, expose from_jsons
Integrating HTTP requests would be antithetical to the project's
security-based nature. Downloading JSONs is better left as an option for
the user, rather than bringing in another library that may have
vulnerabilities.

HostDatabase::from_jsons is now exposed to allow custom JSON files to be
read. This will be integrated into the command-line later.